Make the new module into a very simple web app with Play.
The app can be tested by, from the web module directory, running "sbt" and
then the "~ run" command, then browsing to http://127.0.0.1:9000 .  This means
that sbt is now required for developing in the web module:
  https://github.com/sbt/sbt
  http://www.scala-sbt.org
...but it can be used with just a few sbt commands, like if you type "help" at
the prompt in sbt.

The added application.conf file is a straight copy from the project created
per the link in project/build.properties.