Limit attribute type choices to those used previously.
In other words, make it so that while adding an attribute to an Entity, setting the attribute type lets one choose from only those
Entities that are currently used as attribute types (optionally), instead of having to search through all entities (a time-saver).

Also very minor test improvements and reformatting and refactoring.