Fix some transaction lifetime issues. All tests pass but one, now. Also add a temporary test for exploration purposes -- it will be deleted in the next commit.
The idea in this commit, or what I learned, is that the db layer should
not directly create Entity or similar objects, or call their methods,
but should return IDs and let higher layers create the objects. This and
the fact that the db lifetime should be greater than the transaction
lifetimes, to clean things up and make them work.  And the Entity
etc. lifetime would have to be less than the db (since it contains a db
reference.